+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 * `Djot::Lua` module, which contains ex-`Djot.render_*` methods.
   Now `Djot.render_*` methods are pointers to `Djot::Lua.render_*` (default is not changed for this release).
 
+### Changed
+
+* JavaScript runtime has changed from ExecJS to MiniRacer, for passing Ruby proc to JavaScript runtime.
+  See also [rails/execjs#71 > Custom configuration of runtime](https://github.com/rails/execjs/issues/71).
+
 ## [0.0.5] - 2023-01-21
 
 Note that the default parser might be changed from djot.lua to djot.js in future version.